+++ bmi-calculator/src/components/molecules/ResultCalculator.tsx
@@ -1,0 +1,12 @@
+const ResultCalculator = () => {
+  return (
+    <div className="h-auto w-full space-y-5 rounded-xl bg-blue-500 p-8 text-white">
+      <h4 className="text-preset-4 capitalize">welcome!</h4>
+      <p className="text-preset-7-regular">
+        Enter your height and weight and you'll see your BMI result here
+      </p>
+    </div>
+  );
+};
+
+export default ResultCalculator;
